API for a CatalogObject model of pg_transform
Also adds a new PLJavaBasedLanguage subinterface UsingTransforms,
so a language handler can say implements Routines, UsingTransforms
if it supports applying transforms when executing a routine.

The name UsingTransforms is chosen to allow a possible future
subinterface Transforms, where a language that implements Transforms
could be used to write from-sql and to-sql functions.