Check for UsingTransforms in dispatcher
The dispatcher should reject, at validation and at call time, a routine
declared with TRANSFORM FOR TYPE if the language does not implement
UsingTransforms.

It should call the memo's transforms() method, which will ensure that
any transforms not yet known in the language's transform map are cached
there after checking them with the language's essentialTransformChecks.
Any exception thrown by those checks should happen in this early call
by the dispatcher, so when the language handler itself later calls
transforms(), it need not expect exceptions under normal circumstances.

PostgreSQL, at least through PG 17, does nothing to rule out redundant
TRANSFORM FOR TYPE clauses in routine declarations or resulting
protrftypes arrays with duplicate entries. The dispatcher will check
for that, but only at validation time, as a nudge toward sanity,
but not spend time on that check at call time; in the edge case
of a routine with such a redundant declaration that manages to get
created under conditions that preclude validation, then transforms()
will just return a list with duplicates, which should be benign and
cost less than checking would.