API: separate Memo into Why and How memos
It has become unmistakable that two different kinds of Memo have been
associated with RegProcedure instances. A Why memo pertains to the
intended use of the routine (is it a Validator? a TypeInput function?).
Reflecting the type of Why memo in RegProcedure's type parameter (even
when no actual memo is attached) is useful, providing a compile-time
safeguard against mixing up routines with different purposes.

A How memo pertains to the orthogonal question of how the routine is
implemented. For now, the only two meaningful choices are: implemented
in a PL/Java-based language (and having a PLJavaMemo attached), or
implemented any other way and having no How memo attached.

There is no need for RegProcedure to have a second visible type
parameter, or API to expose a How memo, as it should be uninteresting
(almost everywhere) to know how a routine is implemented. The one place
it is interesting, of course, is in a PL/Java-based language handler.
A PL/Java-based language handler doesn't need a type parameter to know
it's being invoked on a PL/Java-based routine, and the dispatcher can
simply pass the How memo to it as another argument.

The Glot64 example is updated to conform to the updated API.