Better Validator memos, and Handler/InlineHandler
ProceduralLanguageImpl.isPLJavaHandler was applying a ValidatorMemo
to the RegProcedure determined to be the validator of a pljavahandler
language, but no such memo was being applied to the validator of
a PL/Java-based language. That may have simply been an oversight:
it could have been applied (even when memos were sharing one field)
because a PL/Java-based language's validator would never have been
given a PLJavaBased memo (the validator being implemented in the
pljavahandler language, which isn't strictly considered a PLJavaBased
language).

But separating the memos into Why and How makes it easier to reason
about them, and to see what the former Validator memo was incompletely
doing.

A ProceduralLanguage's validator() method points directly to the
RegProcedure representing its validator. The memo on the RegProcedure
is a link in the other direction. That link is one-to-many, as it is
possible to create 'alias' languages (additional language objects
sharing the same handler / inlineHandler / validator functions). The
former ValidatorMemo had just a single field; it now has a Set.

Also, there was only a ValidatorMemo formerly. Now all three of
a language's support functions (Handler, InlineHandler, Validator)
have corresponding memos applied, holding sets of the languages
that have been discovered to rely on them.

The importance of the memo is seen upon invalidation of a support
routine. PostgreSQL will not allow dropping a support routine while
a language depends on it, but will allow CREATE OR REPLACE to change
the routine. PostgreSQL generates an inval event for the changed
support routine but not for the language(s) depending on it. The memo
allows that invalidation to be propagated to invalidate the in-memory
structures for the affected languages, so they will be refreshed from
the catalogs when next needed.