Generalize SupportMemo, start on Transform inval
Move SupportMemo from ProceduralLanguageImpl into RegProcedureImpl and
generalize it so it can be the base of FromSQLMemo and ToSQLMemo for
TransformImpl as well.

Now a Transform will be invalidated if its FromSQL or ToSQL support
function is, and a Transform, when invalidated, will invalidate
RegProcedures that depend on it. A RegProcedure, being invalidated,
removes itself from the dependency lists of any Transforms it had
depended on.

Transform does not (yet) get syscache invalidations. It should, because
current PostgreSQL versions will allow a transform to be dropped
in some circumstances even while routines using it exist.[0][1]

[0] https://www.postgresql.org/message-id/3112950.1743984111%40sss.pgh.pa.us
[1] https://git.postgresql.org/gitweb/?p=postgresql.git;a=commit;h=b73e6d71a89f92436c71467acda1d47552f568a0