Prefer Function for SwitchPointCache initializer
The initializer produced by SwitchPointCache.Builder is something to be
applied to an array of MethodHandle, returning the same array, initialized.

That would make it a Function<MethodHandle[],MethodHandle[]>, or, more
tidily and (you would think) equivalently, UnaryOperator<MethodHandle[]>.

The trouble is, the construction makes heavy use of andThen(...), which
UnaryOperator inherits from Function without refining its return type.
So it's tempting to write andThen(...)::apply, which compiles and works,
and hope javac can simplify it, but it can't, so that in fact bloats
the generated bytecode and the data structure with extra lambdas.

So the bigger mouthful Function<MethodHandle[],MethodHandle[]> is
what it must be.