Quick return in C syscache callbacks if no work
Builds on c13448d to implement the idea from 5adf2c8 to maintain flags
visible from C syscache inval callbacks to indicate whether any instances
of the affected class could require invalidation, and return immediately
in C if not.

This commit handles incrementing an instance count in the "smarter
SwitchPoint" onFirstUse method, and decrementing it in invalidate,
and toggling the C-visible flag when the count is incremented to 1
or decremented to 0.

The case of an unreachable object simply falling out of the CacheMap
is not handled here; if that object had any caching methods called,
the instance count will be left incorrect. This should be a non-issue
for objects encountered in the course of dispatching, as they will be
kept reachable in data structures maintained by the dispatcher. There's
no such assurance, though, of how user code might look up and use catalog
objects.

The actual search of the CacheMap, when not short-circuited, is still
linear. There is no simple improvement using ConcurrentHashMap, which
(even if the hashcode matching PostgreSQL's were to be used) exposes no
API for retrieving entries with a given hashcode. It could be made
logarithmic by using ConcurrentSkipListMap instead; that would make
all the lookups logarithmic too.