Move DualState.SingleSPIfreetuptable free action
This DualState subclass used to free the associated tuple table in the
javaStateUnreachable lifespan event; now, only at javaStateReleased.
It turns out that SPI_freetuptable, since postgres/postgres@3d13623, has
contained code to raise a warning if the tuple table being freed does not
belong to the current SPI connection. With the earlier javaStateUnreachable
handling, that warning could be triggered on rare and irksome occasions
when Java's GC happened to find, during a nested invocation of some Java
function, that a tuple table from an outer invocation had become
unreachable.

It would be conceivable to have javaStateUnreachable try to determine
if the current nest level matches that of the tuple table's creation,
and free it if so at least, otherwise leaking it to the exit of the outer
call. But for now it's also conceivable to just do nothing and let
the context reset at invocation exit mop things up.