API for PL/Java-based set-returning functions
An implementation of PLJavaBasedLanguage may also implement
ReturningSets. If it does, its prepareSRF method, not the usual prepare,
will be used when the target RegProcedure's returnsSet() is true.

The prepareSRF method must return an SRFTemplate. Further, what it
returns must also implement one of SRFTemplate's member subinterfaces,
ValuePerCall or Materialize (PostgreSQL might add additional options in
the future).

The base interface, SRFTemplate, has an abstract 'negotiate' method, to
be passed a list of the subinterfaces the caller is prepared to accept,
and return the index of one from the list that the routine will use.
Each subinterface has a default implementation that will find itself in
the caller's list. A class that implements more than one of the
subinterfaces will inherit conflicting defaults, and therefore have to
provide its own 'negotiate' implementation.

The list of interfaces acceptable to the caller is ordered so as to
reflect the caller's preference. A simple negotiate method could return
the index of the first interface in the list that this SRFTemplate
happens to implement. A more sophisticated one might take properties of
the prepared routine into account.

The ValuePerCall interface specifies a specializeValuePerCall method
that is expected to return an SRFFirst. SRFFirst has a firstCall method
that should return an SRFNext instance. SRFNext has a nextResult method
that will be called as many times as necessary. It should use
fcinfo->result / fcinfo->isNull to store result values for one row, like
any non-set-returning function, but return SINGLE, MULTIPLE, or END to
indicate whether another call is expected. SRFNext implements
AutoCloseable, and its close() method will be called after the last call
made to nextResult (which may happen before nextResult returns END, if
PostgreSQL does not need all the results). The case where nextResult
returns SINGLE is an exception, treated as returning only that one row,
and close() will not be called.

The Materialize interface will likewise specify a specializeMaterialize
method, but the details are TBD, so at this stage of the API the
Materialize interface is a stub and does not specify any usable behavior
yet.