Make DualState.CleanupTracker less restrictive
CleanupTracker is only used when assertions are enabled, and checks
that entries and exits of 'cleanup' loops (cleanEnqueuedInstances for
Java-released or unreachable instances, nativeRelease for expired
native lifespans) are properly paired. It originally prohibited
any more than one cleanup loop being in progress at any time, and
was relaxed in 95d4133 to allow one of each type, assuming there
would be no good reason for such a loop to be reentered.

The generalization in 5a9cdf9 from "resource owner" to Lifespan, with
a possibly-extensible set of Lifespan objects each heading its own list
of dependents, introduced a realistic possibility that some class that
is itself bounded by a Lifespan could also be a Lifespan for other
classes.

A practical application arises in modeling the PostgreSQL ExprContext.
The chief (perhaps only) use of ExprContext in PL/Java is, via its
callback, to signal when no more output is needed from a ValuePerCall
set-returning function that may not have been read to the end. This is
a use of ExprContext as a Lifespan to bound an SRFRoutine.

PostgreSQL's ExprContext. however, does not invoke its callbacks if
error cleanup is afoot, which (ironically enough) would leave its Java
mirror un-cleaned-up in error cases. That can be addressed by also modeling
that ExprContext itself has a lifespan bounded by its per-query memory
context. So ExprContext both has a Lifespan, and is a Lifespan, and in
error cases where the lifespanRelease of the memory context cascades to
lifespanRelease of the ExprContext, the test made here by CleanupTracker
was too restrictive.

The pattern of a Lifespan with a Lifespan does not seem likely to become
so widespread as to cause frequent or deep reentry of lifespanRelease, and
the present example seems a legitimate and reasonable case, so relax the
CleanupTracker assertion to allow it.

In a related change, don't let exceptions abort lifespanRelease:
cleanEnqueuedInstances was already swallowing exceptions (citing JDK 9
Cleaner as precedent) so they would not prevent processing of later
queue entries, but lifespanRelease did not do the same.  Now it does.

It might be better one day to collect exceptions (perhaps as a suppressed
list) to report after the loop.