Model PostgreSQL's ExprContext (internal only)
ExprContextImpl is in the same o.p.p.pg package as implementations of
API-exposed interfaces in o.p.p.model, but there may be no need for any
such API interface, so this is purely for use in the internal module
for now.

For PL/Java's purposes, the chief use of ExprContext is, via its
callback, to be usable as the Lifespan of a DualState instance
associated with the row-collecting activity of a set-returning function
in ValuePerCall mode, whose nativeStateReleased event can trigger
calling the close() method (and resetting internal dispatcher state)
when PostgreSQL collects fewer rows than the function intends. This use
is all internal to the dispatcher.

A PostgreSQL ExprContext carries other information of interest, such as
the per-query memory context, but there is no need to fastidiously
follow PostgreSQL by having the accessor method for that on ExprContext.
As that memory context is needed before the very first call on
a set-returning function, before it is even known whether a Java mirror
of the ExprContext will need to be constructed, it will be better to
fetch that memory context eagerly and put an accessor for it on
ReturnSetInfoImpl instead.

With that in mind, this model class provides no accessor methods at all,
and simply exists to be used as a Lifespan.