Implement set-returning function dispatching (VPC)
Implementation of Materialize mode has to wait, as the API interface
SRFTemplate.Materialize is still only a stub with details TBD.

As contemplated in the implementation of ExprContextImpl, the per-query
memory context is here made available with an accessor method directly
on ReturnSetInfoImpl. It is passed eagerly up by the C dispatch code
through the Java entry points to be readily available, as it will
certainly be needed when any set-returning function is to be called.