Add set-returning functions to Glot64 toy language
As long as the store direction of TupleTableSlot remains unimplemented,
null / void / zero are still the only values any Glot64 function can
return. But now it can return sets of them!

For a Glot64 set-returning functions, there are stricter limits on the
source string. It must be a base64 string that, when 'compiled' (i.e.,
decoded), is the string representation of a decimal integer.  The
function so defined will return that many rows (of null / void / zero),
ignoring any arguments.

If the source string 'compiles' to a negative integer, a single row is
returned, exercising the SRFNext.Result.SINGLE case.