Pick a return type for RegClass.options()
The type List<String> assigned back in 7786fbf already had a comment
that it probably wasn't appropriate, as each String in the list really
represents a key=value pair and that should be made explicit.

On inspection of the PostgreSQL parser and transformRelOptions function,
it's clear the key is an SQL simple identifier, duplicates forbidden,
and the value is a String and never null, so a null-hostile, unmodifiable
Map<Identifier.Simple,String> fits the bill.

An assumption has to be made that the key is never a delimited identifier
with an '=' in it, so the first '=' in the stored value is unambiguously
the delimiter between the key and the value. That's the same assumption
made by PostgreSQL's untransformRelOptions (though, for the time being[0],
PostgreSQL does not reject a key with '=' in it when accepting unknown
custom options such as for foreign data wrappers).

The same Map<Simple,String> type is appropriate also for
Attribute.options() and Attribute.fdwoptions(), and for similar accessors
on foreign-data-related catalogs when those are implemented.

[0] https://www.postgresql.org/message-id/6830EB30.8090904%40acm.org