Use relcache for tuple descriptor in more cases
Getting the descriptor from the typcache is handy when a relation has
an associated type, but not all kinds of relation have one. An index
or TOAST table doesn't, for example.

There was already one case that had to be handled by going to
the relcache, and that was to get the descriptor for pg_class itself,
which can't be expected to find its associated type before it knows
what its columns are. So that code path just needs to be used also
for the relation kinds that don't have an associated type.