API for four new models reachable from RegClass
RegClass should have accessors for AccessMethod and Tablespace
(Database has a Tablespace also).

RegClass indirectly reaches ForeignDataWrapper and ForeignServer.
Opting not to make ForeignTable a CatalogObject in its own right:
it barely qualifies. It is identified by the oid of the RegClass,
and functions more as an extension of that. Its options and
ForeignServer can just be given accessors on RegClass.

Accessors returning these things to be added in a later commit
(along with whitespace-only tidying of lines added here).