Break up DDRProcessor.java
It has grown to such size that breaking it up is long overdue anyway,
but the immediate impetus was to do a trial javadoc run checking
everything (--show-module-contents all --show-packages-all
--show-types private --show-members private). I have no intention of
changing the doc build options to build all that as a matter of course,
but it ought to be possible, so a run with those settings is useful to
identify and fix any javadoc-reported errors that cause the generation
to fail, as opposed to mere warnings.

Javadoc was immediately displeased by the multi-class compilation unit
here.