GetOldestSnapshot -> get_toast_snapshot
When VarlenaWrapper wants to save a TOASTed value that may possibly
be wanted later, it can use the least memory if it just retains the
TOAST pointer and registers a current snapshot in which it is visible.
In PG 9.6, a GetOldestSnapshot function appeared and made that possible.
As of PG 18, with postgres/postgres@4d82750, that function has vanished,
but there is a new one, get_toast_snapshot, that appears to fit the bill.
It's only declared in access/toast_internals.h, though, which isn't
otherwise needed.

Addresses #524.