Added quick sketch of bottom-up FDW java and C
Added quick sketch of bottom-up FDW java and C. It's not remotely functional - but
provides a quick look at what a bottom-up approach may require on the pl/java side.

I assume the C backend will use the 'simple_fdw' I'm developing elsewhere so the
main thing we need on the backend is a way to actually provide a few functions to
the backend and then have the backend make the necessary JNI calls.

One complication is that a wrapper may have multiple servers, each server can have
multiple tables, and each table can have multiple (possibly concurrent) queries. It
would be nice if we could reuse any common wrappers and servers... but for the first
efforts it's probably fine to treat it as one-to-one-to-one as long as we don't
close any doors.