Updated FDW.c
Updated FDW - I had started but got distracted. This is definitely
hand-wavy but it should be good enough to get a good feel on whether
this is a viable approach.

The initial java implementation should be something simple, e.g,
a class with a static List with a few values like (1, 'red'),
(2, 'greem'), etc. That means the only think we need to track is
the current position within the array and to provide the current
row in an acceptable format.