Streamline anachronistic Ptr2Long union usage
PostgreSQL itself has been requiring a C compiler that has uintptr_t
since 9.0 (postgres/postgres@85d02a6). The PL/Java 1.6.x branch only
supports back to PG 9.5, so there is no question the type is supported.

An existing static assert in DualState.c is hereby changed to assert
that sizeof(uintptr_t) <= sizeof(jlong); this is also assured for the
present by a comment upstream in postgres.h section 1: "we require:
sizeof(Datum) == sizeof(void *) == 4 or 8" (where Datum is uintptr_t).

In passing, specifically type as SPIPlanPtr some variables that had
been typed void *. That was once their actual type in the SPI docs,
but not since PG 8.2.

The few sites that ended up as PointerGetDatum(JLongGet(Pointer,...))
are effectively a longwinded way of doing nothing, but I left them
that way to clearly reflect that a pointer is involved, and avoid
assuming PointerGetDatum will always be as simple as it is.