attcacheoff is no longer a pg_attribute member
In PG 18, there is now a CompactAttribute struct found in
tuple descriptors (postgres/postgres@5983a4c) that contains
a field of like purpose, so the one in pg_attribute is gone
(postgres/postgres@02a8d0c). PL/Java deforming wasn't making
any use of it yet anyway.

Where a TupleDesc used to have an attrs offset that was exactly
where a sequence of Form_pg_attribute began, it now has a
compact_attrs offset where a sequence of CompactAttribute starts.
That is still followed by a sequence of Form_pg_attribute, so now
the Java code looking for those has to take the compact_attrs offset
and add natts * sizeof (CompactAttribute).

The CompactAttribute structs were added upstream as a performance
optimization, and could perhaps be made use of here to good effect,
but for now just compute the offset and use the Form_pg_attribute
in the accustomed way. Most promising, perhaps, would be to have
TupleDescriptor make use of the attcacheoff member of the new struct.