Fleshed out FDW a bit.
Fleshed out FDW a bit.

On the pljava-so side it fails to load (in the docker test) because of a dependency on
GLIBC_2.38.  I hadn't hit this with my sister project because it was much simplier and
didn't introduce additional dependencies.

I also found the bits of backend code that lets us access all five types of options at
the points where we need them. It's forcing me to continue to rethink some of my assumptions,
e.g., there's definitely a need to have a conceptual difference between FDW, server, and
table, but it looks like nearly all of the required functions only require a relation or
table.

Could the solution be having the java classes responsible for executing 'CREATE FOREIGN
DATA WRAPPER ...' etc? I had assumed the constructors would need to be called from the
backend - like all of the callbacks - but now think that's backwards. This would makes it
much easier to ensure the options contain the correct classnames and method signatures. :-)

On the java side I have a big chunk of the API and thin implementation written. I haven't
had a chance to introduce some changes I mentioned earlier - ones that will provide a much
better abstraction between the FDW-specific bits and the actual implementation. That was because
I hoped to start seeing immediate feedback as I modified a working (but minimal) FDW.

The Dockerfile under pljava-so had always been a bit of a stopgap measure - a way to do a really
quick sanity check when modifying the backend code - but with the extension additions to the api
and examples modules the docker creation should probably be moved to 'packaging' anyway. At that
point it can continue to use the official postgresql images but use the locally build .so,
jars, etc., instead of trying to sneak in changes to a preconfigured docker image.