Issue 537 - add dockerfile
This patch adds a new profile, 'docker-images', that will create three docker images for each
combination of postgresql version and debian or alpine release. All of the latter are currently
specified in the `pljava_docker/vars/main.yml` file.

The three images are:

- pljava/postgres-pgxnclient - generic pgxnclient built on top of 'postgres' docker image
- pljava/pljava - postgresql server with pl/java preinstalled and configured
- pljava/pljavadev - same as above but with java dev tools added

There is also support for pushing the docker images to a docker repo but that's currently
disabled. It would probably be better to implement pushes as part of a github action.

- ansible

- docker (or podman) server

The ansible dependencies should be pulled in automatically given the metadata in the
`src/main/ansible` directories.

The docker server does not need to be local - there are ways to configure the docker
client to use remote servers.

This can be built locally - but there are good reasons to build and publish the docker
images via github actions.

The details are outside the scope of this patch but it is important to know that it's
an option.

- decide on best way to perform initialization. E.g., should we use an external sql script (e.g., pljava-1.6.9.sql), a separately maintained script, something else?

- implement some 'molecule' tests. They don't have to be very detailed - simply starting a container and running the existing test script should suffice.

- alpine is not supported yet - it uses `apk` instead of `apt` and some of the package names will be different.